feature: add channels 95,96,97,98,99 for use with Rocket M3.
Rocket M3 is a 5.8GHz device with a 2GHz offset transverter.

Channels 95-99 are within the normal spec range of 5.8 GHz Calibration (160MHz wide on chan 100) and should be fine for calibration as it should be part of the normal routine

The band normaly stops before getting to 5470 in most countries (very few have 5470 howver Ubiquiti has supported this on the 5.8GHz gear outside the USA.

Additional testing needing to verify this is indeed sane (Spectrum Analyzer) and possibility for this to go lower on M3 since it is between 2 cal poitns (180MHz appart)

Recommend only using 5MHz width with these.