bugfix: Extended RF channels frequency and channel handleing
Make sure iw/iwinfo display -2/-1 correctly.

Internally -2 is mapped to 254,  -1 to 255

We are relying on a rollover on u8int to make the code change less invasive in iw/iwinfo/mac80211.

Had to modify channel2freq functions to make sure they return correct results.

freq2channel relies on the overflow.